async fix

+++ src/engines/selenium-index.coffee
@@ -669,15 +669,15 @@ module.exports = ->
               if plugin.wdTrace
                 browser.on("status", (info) -> plugin.trace info.cyan)
                 browser.on("command", (meth, path, data) -> plugin.trace "> #{meth.yellow}", path.grey, data || '')
-              if v.browserName in ["chrome","opera"]
+              if v.browserName in ["chrome","opera"] and runner.tests.async.disabled
                 r = browser.init(v).maximize().then(=> promise.call @, browser)
               else
                 r = browser.init(v).then(=> promise.call @, browser)
               browser.qx$browserName = i
               unless binfo.closeBrowser is false or plugin.closeBrowser is (false)
                 r = r.finally =>
-                  unless v.browserName in ["firefox"] then browser.close().quit() else browser.close()
-                  if process.platform[0] is "w"
+                  if (v.browserName not in ["firefox"]) and (runner.tests.async.disabled) then browser.close().quit() else browser.close()
+                  if process.platform[0] is "w" and runner.tests.async.disabled
                     exec("rd /S /Q %temp%")
                     if v.browserName in ["firefox"]
                       exec("taskkill /F /T /IM firefox.exe")
